@@ -8,7 +8,7 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.2.5",
+        "php": "^7.3|^8.0",
         "arandilopez/laravel-profane": "^0.5.0",
         "artesaos/seotools": "^0.19.0",
         "backpack/crud": "4.1.*",
@@ -17,38 +17,38 @@
         "christiankuri/laravel-favorite": "^1.4",
         "consoletvs/charts": "6.*",
         "creativeorange/gravatar": "^1.0",
-        "fideloper/proxy": "^4.2",
-        "fruitcake/laravel-cors": "^1.0",
-        "guzzlehttp/guzzle": "^6.3",
+        "fideloper/proxy": "^4.4",
+        "fruitcake/laravel-cors": "^2.0",
+        "guzzlehttp/guzzle": "^7.0.1",
         "kyslik/column-sortable": "^6.3",
-        "laravel/framework": "^7.0",
-        "laravel/passport": "^9.2",
-        "laravel/tinker": "^2.0",
-        "laravel/ui": "^2.0",
-        "laravelcollective/html": "^6.1",
+        "laravel/framework": "^8.45",
+        "laravel/passport": "^10.0",
+        "laravel/tinker": "^2.5",
+        "laravel/ui": "^3.2",
+        "laravelcollective/html": "^6.2",
         "league/flysystem-aws-s3-v3": "^1.0",
         "mll-lab/graphql-php-scalars": "^3.1",
         "mll-lab/laravel-graphql-playground": "^2.1",
-        "nagy/laravel-rating": "^1.4",
-        "nuwave/lighthouse": "^4.14",
-        "rollbar/rollbar-laravel": "^6.0",
-        "spatie/laravel-backup": "^6.11",
+        "nagy/laravel-rating": "^2.0",
+        "nuwave/lighthouse": "^5.12",
+        "rollbar/rollbar-laravel": "^7.0",
+        "spatie/laravel-backup": "^6.16",
         "tucker-eric/eloquentfilter": "^2.3"
     },
     "require-dev": {
         "backpack/generators": "^3.1",
-        "barryvdh/laravel-debugbar": "^3.4",
-        "barryvdh/laravel-ide-helper": "^2.8",
-        "facade/ignition": "^2.0",
-        "fzaninotto/faker": "^1.9.1",
+        "barryvdh/laravel-debugbar": "^3.6",
+        "barryvdh/laravel-ide-helper": "^2.10",
+        "facade/ignition": "^2.5",
         "haydenpierce/class-finder": "^0.4.2",
-        "laracasts/generators": "^1.1",
-        "laravel-frontend-presets/tailwindcss": "^4.1",
+        "laracasts/generators": "^2.0",
+        "laravel-frontend-presets/tailwindcss": "^4.3",
         "laravel/homestead": "^10.8",
-        "matt-allan/laravel-code-style": "^0.5.1",
-        "mockery/mockery": "^1.3.1",
-        "nunomaduro/collision": "^4.1",
-        "phpunit/phpunit": "^8.5"
+        "matt-allan/laravel-code-style": "^0.6",
+        "mockery/mockery": "^1.4.2",
+        "nunomaduro/collision": "^5.0",
+        "phpunit/phpunit": "^9.3.3",
+        "fakerphp/faker": "^1.9.1"
     },
     "config": {
         "optimize-autoloader": true,
@@ -62,12 +62,10 @@
     },
     "autoload": {
         "psr-4": {
-            "App\\": "app/"
-        },
-        "classmap": [
-            "database/seeds",
-            "database/factories"
-        ]
+            "App\\": "app/",
+            "Database\\Factories\\": "database/factories/",
+            "Database\\Seeders\\": "database/seeders/"
+        }
     },
     "autoload-dev": {
         "psr-4": {
