@@ -4,6 +4,7 @@ namespace App;
 
 use Backpack\CRUD\app\Models\Traits\CrudTrait;
 use EloquentFilter\Filterable;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
 /**
@@ -36,10 +37,10 @@ use Illuminate\Database\Eloquent\Model;
  */
 class Patch extends Model
 {
+    use HasFactory;
     use Filterable, CrudTrait;
-
-    protected $dates = [
-        'launched_at',
+    protected $casts = [
+        'launched_at' => 'datetime',
     ];
 
     protected $guarded = [];
