@@ -4,6 +4,7 @@ namespace App;
 
 use Backpack\CRUD\app\Models\Traits\CrudTrait;
 use Carbon\Carbon;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 
 /**
@@ -32,12 +33,12 @@ use Illuminate\Database\Eloquent\Model;
  */
 class Post extends Model
 {
+    use HasFactory;
     use CrudTrait;
 
     protected $guarded = [];
-
-    protected $dates = [
-        'published_at',
+    protected $casts = [
+        'published_at' => 'datetime',
     ];
 
     public function user()
