@@ -1,5 +1,7 @@
 <?php
 
+namespace Database\Seeders;
+
 use App\Post;
 use Illuminate\Database\Seeder;
 
@@ -7,6 +9,6 @@ class PostSeeder extends Seeder
 {
     public function run()
     {
-        factory(Post::class, 20)->create();
+        Post::factory()->count(20)->create();
     }
 }
