@@ -1,5 +1,12 @@
 <?php
 
+use App\Http\Controllers\BuildMetricController;
+use App\Http\Controllers\DashboardController;
+use App\Http\Controllers\LoadoutsController;
+use App\Http\Controllers\PostController;
+use App\Http\Controllers\ProfileController;
+use App\Http\Controllers\SettingsController;
+use App\Http\Controllers\SubscriptionController;
 use Illuminate\Support\Facades\Route;
 
 /*
@@ -14,33 +21,33 @@ use Illuminate\Support\Facades\Route;
 */
 
 Route::middleware(['auth'])->group(function () {
-    Route::get('profile/{id}/edit', 'ProfileController@editProfile')->where('id', '[0-9]+');
+    Route::get('profile/{id}/edit', [ProfileController::class, 'editProfile'])->where('id', '[0-9]+');
 
     // TODO: need to review this
-    Route::post('profile/{id}/edit/update', 'ProfileController@editProfileSave')
+    Route::post('profile/{id}/edit/update', [ProfileController::class, 'editProfileSave'])
         ->name('user.profile.update')
         ->where('id', '[0-9]+');
 
-    Route::delete('build/delete/{id}', 'LoadoutsController@delete');
+    Route::delete('build/delete/{id}', [LoadoutsController::class, 'delete']);
 });
 
 Route::middleware(['can:access-api'])->group(function () {
-    Route::get('settings/tokens', 'SettingsController@tokens')->name('settings.tokens');
+    Route::get('settings/tokens', [SettingsController::class, 'tokens'])->name('settings.tokens');
 });
 
 // Public, signed unsubscribe url
-Route::get('unsubscribe/{user}', 'SubscriptionController@unsubscribe')
+Route::get('unsubscribe/{user}', [SubscriptionController::class, 'unsubscribe'])
     ->name('unsubscribe')
     ->middleware('signed');
 
-Route::get('/', 'DashboardController@index');
-Route::get('browse', 'LoadoutsController@index')->name('loadout.index');
-Route::resource('blog', 'PostController')->only(['index', 'show']);
+Route::get('/', [DashboardController::class, 'index']);
+Route::get('browse', [LoadoutsController::class, 'index'])->name('loadout.index');
+Route::resource('blog', PostController::class)->only(['index', 'show']);
 
-Route::get('preview/{loadoutId}', 'LoadoutsController@preview')->name('preview.show');
-Route::get('build', 'LoadoutsController@build');
-Route::get('build/{loadoutId}', 'LoadoutsController@build')->name('build.show');
+Route::get('preview/{loadoutId}', [LoadoutsController::class, 'preview'])->name('preview.show');
+Route::get('build', [LoadoutsController::class, 'build']);
+Route::get('build/{loadoutId}', [LoadoutsController::class, 'build'])->name('build.show');
 Route::view('/privacy-policy', 'privacy-policy.index');
-Route::get('asv/{gun}/{combo}', 'BuildMetricController@show')->name('asv.show')->where('gun', '[0-9]+');
-Route::get('profile/{id}', 'ProfileController@index')->name('user.profile')->where('id', '[0-9]+');
+Route::get('asv/{gun}/{combo}', [BuildMetricController::class, 'show'])->name('asv.show')->where('gun', '[0-9]+');
+Route::get('profile/{id}', [ProfileController::class, 'index'])->name('user.profile')->where('id', '[0-9]+');
 Auth::routes();
