@@ -16,17 +16,17 @@ class LoadoutScopesTest extends TestCase
     public function testOnOlderPatchScopeGrabsOldLoadouts()
     {
         // Order here matters because of our onSaved model event!
-        $oldPatch = factory(Patch::class)->create([
+        $oldPatch = Patch::factory()->create([
             'launched_at' => Carbon::now()->subYear(),
         ]);
-        $oldLoadouts = factory(Loadout::class, 3)->create([
+        $oldLoadouts = Loadout::factory()->count(3)->create([
             'patch_id' => '1',
         ]);
 
-        $newPatch = factory(Patch::class)->create([
+        $newPatch = Patch::factory()->create([
             'launched_at' => Carbon::now(),
         ]);
-        $newLoadout = factory(Loadout::class, 2)->create([
+        $newLoadout = Loadout::factory()->count(2)->create([
             'patch_id' => $newPatch->id,
         ]);
 
