@@ -19,7 +19,7 @@ class OverclockCrudController extends CrudController
 
     public function setup()
     {
-        $this->crud->setModel('App\Overclock');
+        $this->crud->setModel(\App\Overclock::class);
         $this->crud->setRoute(config('backpack.base.route_prefix').'/overclock');
         $this->crud->setEntityNameStrings('overclock', 'overclocks');
     }
@@ -110,7 +110,7 @@ class OverclockCrudController extends CrudController
             'name' => 'character_id', // the relationship name in your Model
             'entity' => 'character', // the relationship name in your Model
             'attribute' => 'name', // attribute on Article that is shown to admin
-            'model' => "App\Character",
+            'model' => \App\Character::class,
             'tab' => 'Base Info',
         ]);
         $this->crud->addField([
@@ -118,7 +118,7 @@ class OverclockCrudController extends CrudController
             'name' => 'gun_id', // the relationship name in your Model
             'entity' => 'gun', // the relationship name in your Model
             'attribute' => 'name', // attribute on Article that is shown to admin
-            'model' => "App\Gun",
+            'model' => \App\Gun::class,
             'tab' => 'Base Info',
         ]);
         $this->crud->addField([   // select_from_array
@@ -220,7 +220,7 @@ class OverclockCrudController extends CrudController
             'name' => 'character_id', // the relationship name in your Model
             'entity' => 'character', // the relationship name in your Model
             'attribute' => 'name', // attribute on Article that is shown to admin
-            'model' => "App\Character",
+            'model' => \App\Character::class,
             'tab' => 'Base Info',
         ]);
         $this->crud->addField([
@@ -228,7 +228,7 @@ class OverclockCrudController extends CrudController
             'name' => 'gun_id', // the relationship name in your Model
             'entity' => 'gun', // the relationship name in your Model
             'attribute' => 'name', // attribute on Article that is shown to admin
-            'model' => "App\Gun",
+            'model' => \App\Gun::class,
             'tab' => 'Base Info',
         ]);
 
