@@ -1,14 +1,30 @@
 <?php
 
-/** @var \Illuminate\Database\Eloquent\Factory $factory */
+namespace Database\Factories;
 
 use App\Build;
-use Faker\Generator as Faker;
+use Illuminate\Database\Eloquent\Factories\Factory;
 
-$factory->define(Build::class, function (Faker $faker) {
-    return [
-        'name' => 'Super Awesome Build',
-        'user_id' => 1,
-        'character_id' => 1,
-    ];
-});
+class BuildFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = Build::class;
+
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition()
+    {
+        return [
+            'name' => 'Super Awesome Build',
+            'user_id' => 1,
+            'character_id' => 1,
+        ];
+    }
+}
