@@ -1,21 +1,37 @@
 <?php
 
-/** @var \Illuminate\Database\Eloquent\Factory $factory */
+namespace Database\Factories;
 
 use App\Loadout;
-use Faker\Generator as Faker;
+use Illuminate\Database\Eloquent\Factories\Factory;
 
-$factory->define(Loadout::class, function (Faker $faker) {
-    $createdAt = $faker->dateTimeThisYear;
+class LoadoutFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = Loadout::class;
 
-    return [
-        'name' => $faker->name,
-        'description' => $faker->sentence,
-        'user_id' => 1,
-        'character_id' => $faker->numberBetween(1, 4),
-        'patch_id' => 1,
-        'created_at' => $createdAt,
-        'updated_at' => $createdAt,
-        'throwable_id' => 12,
-    ];
-});
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition()
+    {
+        $createdAt = $this->faker->dateTimeThisYear;
+
+        return [
+            'name' => $this->faker->name,
+            'description' => $this->faker->sentence,
+            'user_id' => 1,
+            'character_id' => $this->faker->numberBetween(1, 4),
+            'patch_id' => 1,
+            'created_at' => $createdAt,
+            'updated_at' => $createdAt,
+            'throwable_id' => 12,
+        ];
+    }
+}
