@@ -1,17 +1,33 @@
 <?php
 
-/** @var \Illuminate\Database\Eloquent\Factory $factory */
+namespace Database\Factories;
 
 use App\Mod;
-use Faker\Generator as Faker;
+use Illuminate\Database\Eloquent\Factories\Factory;
 
-$factory->define(Mod::class, function (Faker $faker) {
-    return [
-        'name' => $faker->name,
-        'effect' => '+1.0 Rate of Fire',
-        'description' => $faker->sentence,
-        'row' => 1,
-        'column' => 1,
-        'gun_id' => 1,
-    ];
-});
+class ModFactory extends Factory
+{
+    /**
+     * The name of the factory's corresponding model.
+     *
+     * @var string
+     */
+    protected $model = Mod::class;
+
+    /**
+     * Define the model's default state.
+     *
+     * @return array
+     */
+    public function definition()
+    {
+        return [
+            'name' => $this->faker->name,
+            'effect' => '+1.0 Rate of Fire',
+            'description' => $this->faker->sentence,
+            'row' => 1,
+            'column' => 1,
+            'gun_id' => 1,
+        ];
+    }
+}
