@@ -1,5 +1,10 @@
 <?php
 
+use App\Http\Controllers\Admin\Charts\DailyLoadoutsChartController;
+use App\Http\Controllers\Admin\Charts\DailyUpdatedLoadoutsChartController;
+use App\Http\Controllers\Admin\Charts\DailyUsersChartController;
+use Illuminate\Support\Facades\Route;
+
 // --------------------------
 // Custom Backpack Routes
 // --------------------------
@@ -23,9 +28,9 @@ Route::group([
     Route::crud('character', 'CharacterCrudController');
     Route::crud('post', 'PostCrudController');
 
-    Route::get('charts/daily-users', 'Charts\DailyUsersChartController@response')->name('charts.daily-users.index');
-    Route::get('charts/daily-loadouts', 'Charts\DailyLoadoutsChartController@response')->name('charts.daily-loadouts.index');
-    Route::get('charts/daily-updated-loadouts', 'Charts\DailyUpdatedLoadoutsChartController@response')->name('charts.daily-updated-loadouts.index');
+    Route::get('charts/daily-users', [DailyUsersChartController::class, 'response'])->name('charts.daily-users.index');
+    Route::get('charts/daily-loadouts', [DailyLoadoutsChartController::class, 'response'])->name('charts.daily-loadouts.index');
+    Route::get('charts/daily-updated-loadouts', [DailyUpdatedLoadoutsChartController::class, 'response'])->name('charts.daily-updated-loadouts.index');
 }); // this should be the absolute last line of this file
 
 Route::group([
